test: setup zig and ensure tests run sequentially
