+++ package.json
@@ -36,8 +36,7 @@
     "ts-jest": "^29.3.0",
     "ts-node": "^10.9.2",
     "typescript": "^5.8.2",
-    "typescript-eslint": "^8.28.0",
-    "@ziglang/cli": "^0.0.14"
+    "typescript-eslint": "^8.28.0"
   },
   "dependencies": {
     "@orcabus/platform-cdk-constructs": "^0.0.9",
